+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 ### 🐣新特性
 * 【system 】     OshiUtil增加getCurrentProcess方法
 * 【extra  】     SpringUtil增加getApplicationName、publishEvent方法（issue#I485NZ@Gitee）
+* 【core   】     BeanUtil.getProperty增加判空（issue#I488HA@Gitee）
 
 ### 🐞Bug修复